add ui struct with run method
