+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod cell;
 pub mod grid;
+pub mod ui;
